Web/来自一年后的告白回复/requirements.txt
Flask==3.0.3
pycryptodome==3.21.0
Requests==2.32.3
sympy==1.12
Werkzeug==3.0.4
